fix: tests for statuses
